move repositiationTouchSelection to Touch plugin
