+++ packages/roosterjs-content-model-plugins/lib/touch/TouchPlugin.ts
@@ -1,7 +1,14 @@
-import type { EditorPlugin, IEditor, PluginEvent } from 'roosterjs-content-model-types';
+import type {
+    EditorPlugin,
+    IEditor,
+    PluginEvent,
+    ContentModelText,
+} from 'roosterjs-content-model-types';
 import { getNodePositionFromEvent } from '../utils/getNodePositionFromEvent';
-import { repositionTouchSelection } from './repositionTouchSelection';
+import { getSelectedSegmentsAndParagraphs } from 'roosterjs-content-model-dom';
+import { adjustWordSelection } from 'roosterjs-content-model-api';
 
+const MAX_TOUCH_MOVE_DISTANCE = 6; // the max number of offsets for the touch selection to move
 const POINTER_DETECTION_DELAY = 150; // Delay time to wait for selection to be updated and also detect if pointerup is a tap or part of double tap
 const PUNCTUATION_MATCHING_REGEX = /[.,;:!]/;
 const SPACE_MATCHING_REGEX = /\s/;
@@ -73,7 +80,7 @@ export class TouchPlugin implements EditorPlugin {
 
                     if (this.editor) {
                         if (!this.isDblClicked) {
-                            repositionTouchSelection(this.editor);
+                            this.repositionTouchSelection();
 
                             // reset values
                             this.isTouchPenPointerEvent = false;
@@ -165,6 +172,83 @@ export class TouchPlugin implements EditorPlugin {
                 break;
         }
     }
+
+    repositionTouchSelection = () => {
+        this.editor?.formatContentModel(
+            (model, _context) => {
+                if (this.editor) {
+                    let segmentAndParagraphs = getSelectedSegmentsAndParagraphs(
+                        model,
+                        false /*includingFormatHolder*/,
+                        true /*includingEntity*/,
+                        true /*mutate*/
+                    );
+
+                    const isCollapsedSelection =
+                        segmentAndParagraphs.length >= 1 &&
+                        segmentAndParagraphs.every(x => x[0].segmentType == 'SelectionMarker');
+
+                    // 1. adjust selection to a word if selection is collapsed
+                    if (isCollapsedSelection) {
+                        const para = segmentAndParagraphs[0][1];
+                        const path = segmentAndParagraphs[0][2];
+
+                        segmentAndParagraphs = adjustWordSelection(
+                            model,
+                            segmentAndParagraphs[0][0]
+                        ).map(x => [x, para, path]);
+
+                        // 2. Collect all text segments in selection
+                        const segments: ContentModelText[] = [];
+                        segmentAndParagraphs.forEach(item => {
+                            if (item[0].segmentType == 'Text') {
+                                segments.push(item[0]);
+                            }
+                        });
+
+                        // If there are 3 text segment in the Word, selection is in middle of the word
+                        // before selection marker + after selection marker
+                        if (segments.length === 2) {
+                            // 3. Calculate the offset to move cursor to the nearest edge of the word if within 6 characters
+                            // default to end of the word if user tapped in the middle
+                            const leftCursorWordLength = segments[0].text.length;
+                            const rightCursorWordLength = segments[1].text.length;
+                            let movingOffset: number =
+                                leftCursorWordLength >= rightCursorWordLength
+                                    ? rightCursorWordLength
+                                    : -leftCursorWordLength;
+                            movingOffset =
+                                Math.abs(movingOffset) > MAX_TOUCH_MOVE_DISTANCE ? 0 : movingOffset;
+
+                            // 4. Move cursor to the calculated offset
+                            const selection = this.editor.getDOMSelection();
+                            if (selection?.type == 'range' && movingOffset !== 0) {
+                                const selectedRange = selection.range;
+                                const newRange = this.editor.getDocument().createRange();
+                                newRange.setStart(
+                                    selectedRange.startContainer,
+                                    selectedRange.startOffset + movingOffset
+                                );
+                                newRange.setEnd(
+                                    selectedRange.endContainer,
+                                    selectedRange.endOffset + movingOffset
+                                );
+                                this.editor.setDOMSelection({
+                                    type: 'range',
+                                    range: newRange,
+                                    isReverted: false,
+                                });
+                            }
+                        }
+                    }
+                }
+                return false;
+            },
+            {
+                apiName: 'TouchSelection',
+            }
+        );
+    };
 }
 
 /**
